Content Helper: Rename some ts/tsx files to resemble more to WP standards
